api_yamdb/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
class-registry==2.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==2.2.16
django-cors-headers==3.11.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
filters==1.3.2
flake8==4.0.1
gunicorn==20.0.4
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.3
pluggy==0.13.1
#psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
regex==2022.3.15
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
#social-auth-app-django==4.0.0
#social-auth-core==4.2.0
sqlparse==0.3.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
